widgets: Fix support for smooth scrolling in gradient editor
Smooth scrolling requires that we keep the scroll (or zoom, as is in
this case) state as a floating-point number. The scroll deltas are often
less than one, thus storing scroll state as an integer will effectively
ignore most of the scroll events.

+++ app/widgets/gimpgradienteditor.h
@@ -60,7 +60,7 @@ struct _GimpGradientEditor
   GtkWidget              *control;
 
   /*  Zoom and scrollbar  */
-  guint                   zoom_factor;
+  gdouble                 zoom_factor;
   GtkAdjustment          *scroll_data;
 
   /*  Gradient view  */
